pyproject.toml
[project]
name = "django-bolt-with-batteries"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asgiref==3.11.0",
    "asttokens==3.0.1",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "decorator==5.2.1",
    "django==6.0.1",
    "django-bolt",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-yasg==1.21.11",
    "executing==2.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "inflection==0.5.1",
    "iniconfig==2.3.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "matplotlib-inline==0.2.1",
    "maturin==1.11.5",
    "msgspec==0.20.0",
    "packaging==25.0",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.52",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-django==4.11.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "ruff==0.14.11",
    "sqlparse==0.5.5",
    "stack-data==0.6.3",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uv==0.9.26",
    "uvloop==0.22.1",
    "wcwidth==0.2.14",
]

[tool.uv.sources]
django-bolt = { git = "https://github.com/Manav1011/django-bolt" }
